Merge many and manyRange into a single function that takes options (#23)
This change also allows many to "collect" the results into an arraylist
and return a slice of results instead of just the parsed string. To get
the old behavior, one can still set `collect` to `false`. many and manyN
now both also take a separator parse, for parsing things that separates
elements. This function defaults to noop.